backend/requirements.txt
# Core AI and ML libraries
agno>=0.1.0
groq>=0.4.0

# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Enhanced monitoring and logging
python-json-logger>=2.0.0
structlog>=23.0.0

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.0
pypdf>=3.17.0

# Database and storage
sqlalchemy>=2.0.0

# HTTP and API requests
requests>=2.31.0
aiohttp>=3.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Enhanced utilities
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Legal research specific
duckduckgo-search>=4.1.0

# Async processing
aiofiles>=23.2.0

# Monitoring and metrics
prometheus-client>=0.19.0
psutil>=5.9.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Security
cryptography>=41.0.0

# File handling
python-magic>=0.4.27

# Enhanced error handling
tenacity>=8.2.0

# Performance monitoring
memory-profiler>=0.61.0 